edit print_binary file

+++ print_binary.c
@@ -2,7 +2,7 @@
 
 #define bit32_range (4294967296)
 
-void print_binary(long long int number, int *count);
+void print_binary(long int number, int *count);
 
 /**
  * print_positive_binary - print the binary representation of a positive number
@@ -10,10 +10,11 @@ void print_binary(long long int number, int *count);
  * @count: a pointer to the count of printed characters
  * Return: void
  */
-void print_positive_binary(long long int number, int *count)
+
+void print_positive_binary(long int number, int *count)
 {
-	long long int temp = 0;
-	long long int max = 1;
+	long int temp = 0;
+	long int max = 1;
 
 	if (number == 0)
 	{
@@ -48,12 +49,14 @@ void print_positive_binary(long long int number, int *count)
 	}
 }
 }
+
 /**
  * print_binary - print the binary representation of a number
  * @number: the number to print
  * @count: a pointer to the count of printed characters
  */
-void print_binary(long long int number, int *count)
+
+void print_binary(long int number, int *count)
 {
 	if (number < 0)
 	{
